Move files out of SpaceDashboard into /Space or /AccountDropdown folder
