+++ ui/src/AccountDropdown/InviteContributorsConfirmationModal.scss
