+++ ui/src/Application/PeopleMover.tsx
@@ -40,8 +40,8 @@ import {Person} from '../People/Person';
 import PeopleClient from '../People/PeopleClient';
 import Header from '../Header/Header';
 import {Redirect} from 'react-router-dom';
-import {Space} from '../SpaceDashboard/Space';
-import SpaceClient from '../SpaceDashboard/SpaceClient';
+import {Space} from '../Space/Space';
+import SpaceClient from '../Space/SpaceClient';
 import {Product} from '../Products/Product';
 import ReassignedDrawer from '../ReassignedDrawer/ReassignedDrawer';
 import {ProductTag} from '../ProductTag/ProductTag';
