+++ ui/src/SpaceDashboard/SpaceForm.tsx
@@ -18,8 +18,8 @@
 import React, {FormEvent, useState} from 'react';
 import {closeModalAction, fetchUserSpacesAction} from '../Redux/Actions';
 import {connect} from 'react-redux';
-import SpaceClient from './SpaceClient';
-import {createEmptySpace, Space} from './Space';
+import SpaceClient from '../Space/SpaceClient';
+import {createEmptySpace, Space} from '../Space/Space';
 import FormButton from '../ModalFormComponents/FormButton';
 
 import './SpaceForm.scss';
